requirements.txt
opencv-python
pandas
google-generativeai
Pillow
numpy
torch
torchvision
roboflow
python-dotenv
supervision
tqdm
moviepy
ultralytics
rfdetr
python-dotenv